requirements.txt
Django==1.7.7
django-appconf==1.0.1
django-user-accounts==1.0.1
djangorestframework==3.1.1
oauthlib==0.7.2
Pillow==2.7.0
psycopg2==2.6
pytz==2015.2
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0